+++ MovieQuiz/Models/QuizStepViewModel.swift
@@ -1,0 +1,8 @@
+import UIKit
+import Foundation
+
+struct QuizStepViewModel {
+    let image: UIImage
+    let question: String
+    let questionNumber: String
+}
